refactor apollo
